backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.6.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.0
beautifulsoup4>=4.12.0
requests>=2.31.0
lxml>=5.0.0
cloudscraper>=1.2.70